Enabled warnings=errors for GCC and then made things compile. Had to turn off strict aliasing so we could type-pun the function pointers. I think there's a better solution there, but we're stricter than we were.

+++ libmemcached/memcached_analyze.c
@@ -27,7 +27,7 @@ static void calc_least_free_node(memcached_analysis_st *result,
                                  const long max_allowed_bytes,
                                  const long used_bytes)
 {
-  long remaining_bytes= max_allowed_bytes - used_bytes;
+  uint64_t remaining_bytes= (uint64_t)max_allowed_bytes - used_bytes;
 
   if (result->least_remaining_bytes == 0 ||
       remaining_bytes < result->least_remaining_bytes)
